backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn>=0.15.0,<0.16.0
motor==3.3.1
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart>=0.0.5
python-dotenv==1.0.0
pymongo==4.6.0
email-validator==2.1.0.post1
loguru==0.7.2

# Database dependencies
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
alembic==1.13.1

# AI/ML dependencies
google-generativeai==0.3.1
numpy==1.26.2
scikit-learn==1.3.2

# Background tasks
celery==5.3.6
flower==2.0.1

# HTTP client
aiohttp==3.9.1
httpx==0.25.1

# XML parsing
xmltodict==0.13.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1
pre-commit==3.5.0

# New dependencies
watchfiles>=0.18.0
websockets>=10.4
cloudinary>=1.42.0
